% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{ollama_check}
\alias{ollama_check}
\title{Check that ollama is installed with required models, and download if not.}
\usage{
ollama_check(sudo = is_docker_sudo())
}
\arguments{
\item{sudo}{Set to \code{TRUE} if ollama is running in docker with sudo
privileges.}
}
\value{
TRUE if everything works okay, otherwise the function will error
before returning.
}
\description{
Check that ollama is installed with required models, and download if not.
}
\examples{
\dontrun{
chk <- ollama_check ()
}

}
\concept{ollama}
