% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{pkgmatch_load_data}
\alias{pkgmatch_load_data}
\title{Load 'pkgmatch' data for specified corpus.}
\usage{
pkgmatch_load_data(what = "idfs", corpus = "ropensci", fns = FALSE)
}
\arguments{
\item{what}{One of the three data types described above: "idfs",
"functions", or "calls".}

\item{corpus}{Must be specified as one of "ropensci", "cran", or "bioc" (for
BioConductor). If \code{idfs} parameter is not specified, data will be
automatically downloaded for the corpus specified by this parameter. The
function will then return the most similar package from the specified
corpus. Note that calculations will \code{corpus = "cran"} will generally take
longer, because the corpus is much larger.}

\item{fns}{If \code{FALSE} (default), load data for all packages; otherwise
load (considerably larger dataset of) data for all individual
functions.}
}
\value{
The loaded data.
}
\description{
Load pre-computed data for a specified corpus. Data types are:
\itemize{
\item "idfs" for Inverse Document Frequency weightings;
\item "functions" for frequency tables for text descriptions of function
calls; or
\item "calls" for frequency tables for actual function calls.
}

This function is called within the main \link{pkgmatch_similar_pkgs}
function to load required data there, and should not generally need to be
explicitly called.
}
\examples{
corpus <- "cran"
generate_pkgmatch_example_data (corpus = corpus)
idfs <- pkgmatch_load_data ("idfs", corpus = corpus)
idfs_fns <- pkgmatch_load_data ("idfs", fns = TRUE, corpus = corpus)
}
\seealso{
Other utils:
\code{\link[=generate_pkgmatch_example_data]{generate_pkgmatch_example_data()}},
\code{\link[=head.pkgmatch]{head.pkgmatch()}},
\code{\link[=pkgmatch_browse]{pkgmatch_browse()}},
\code{\link[=pkgmatch_update_cache]{pkgmatch_update_cache()}},
\code{\link[=print.pkgmatch]{print.pkgmatch()}}
}
\concept{utils}
