% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sitter.R
\name{pkgmatch_treesitter_fn_tags}
\alias{pkgmatch_treesitter_fn_tags}
\title{Identify all function calls make within a package.}
\usage{
pkgmatch_treesitter_fn_tags(path)
}
\arguments{
\item{path}{Path to local package, or \code{.tar.gz} file of package source.}
}
\value{
A \code{data.frame} of all function calls made within the package, with
the following columns:
\itemize{
\item 'fn' Name of the package function within which call is made, including
namespace identifiers of "::" for exported functions and ":::" for
non-exported functions.
\item name Name of function being called, including namespace.
\item start Byte number within file corresponding to start of definition
\item end Byte number within file corresponding to end of definition
\item file Name of file in which fn call is defined.
}
}
\description{
This function uses "treesitter"
(\url{https://github.com/tree-sitter/tree-sitter}) to tag all function calls
made within a local package, and to associate those calls with package
namespaces.

This is used as input to the \link{pkgmatch_bm25_fn_calls} function, to
enable function calls within a local package to be inversely weighted by
frequencies within all packages within a corpus. The results of applying
this function to the full corpora used in this package are contained within
the data listed on
\url{https://github.com/ropensci-review-tools/pkgmatch/releases/tag/v0.4.0},
as "fn-calls-ropensci.Rds" and "fn-calls-cran.Rds".
}
\examples{
# Get function calls made within locally-installed packages:
fn_tags <- pkgmatch_treesitter_fn_tags ("curl") # Name of installed package
fn_tags <- pkgmatch_treesitter_fn_tags ("cli") # Name of installed package

# Or get calls from full source code:
u <- "https://cran.r-project.org/src/contrib/odbc_1.5.0.tar.gz"
path <- file.path (tempdir (), basename (u))
\dontrun{
download.file (u, destfile = path)
fn_tags <- pkgmatch_treesitter_fn_tags (path)
}
}
\concept{treesitter}
