% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgmatch-methods.R
\name{print.pkgmatch}
\alias{print.pkgmatch}
\title{Print method for 'pkgmatch' objects}
\usage{
\method{print}{pkgmatch}(x, ...)
}
\arguments{
\item{x}{Object to be printed}

\item{...}{Additional parameters passed to default 'print' method.}
}
\value{
The result of printing \code{x}, in form of either a single character
vector, or a named list of character vectors.
}
\description{
The main \code{pkgmatch} function, \link{pkgmatch_similar_pkgs},
returns \code{data.frame} objects of class "pkgmatch". This class exists
primarily to enable this print method, which summarises by default the top 5
matching packages or functions. Objects can be converted to standard
\code{data.frame}s with \code{as.data.frame()}.
}
\examples{
corpus <- "cran"
set.seed (1L)
generate_pkgmatch_example_data (corpus = corpus)
input <- "Download open spatial data from NASA"
p <- pkgmatch_similar_pkgs (input, corpus = corpus)
head (p) # Shows first 5 rows of full `data.frame` object
p # Default print method, lists 5 best matching packages
}
\seealso{
Other utils:
\code{\link[=generate_pkgmatch_example_data]{generate_pkgmatch_example_data()}},
\code{\link[=head.pkgmatch]{head.pkgmatch()}},
\code{\link[=pkgmatch_browse]{pkgmatch_browse()}},
\code{\link[=pkgmatch_load_data]{pkgmatch_load_data()}},
\code{\link[=pkgmatch_update_cache]{pkgmatch_update_cache()}}
}
\concept{utils}
