% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{serve_api}
\alias{serve_api}
\title{serve plumber API to interface with ollama server}
\usage{
serve_api(port = 8000L)
}
\arguments{
\item{port}{Port for API to be exposed on}
}
\value{
Nothing; calling this starts a blocking process.
}
\description{
serve plumber API to interface with ollama server
}
\concept{api}
