% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_ollama_url}
\alias{set_ollama_url}
\title{Set the URL for local ollama API}
\usage{
set_ollama_url(ollama_url)
}
\arguments{
\item{ollama_url}{The desired ollama API URL}
}
\value{
The ollama API URL
}
\description{
Set the URL for local ollama API
}
\seealso{
\code{\link[=get_ollama_url]{get_ollama_url()}}

Other ollama: 
\code{\link{get_ollama_url}()},
\code{\link{ollama_check}()}
}
\concept{ollama}
