% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{generate_pkgmatch_example_data}
\alias{generate_pkgmatch_example_data}
\title{Generate example data to use with pkgmatch}
\usage{
generate_pkgmatch_example_data()
}
\value{
(Invisibly) The path to the temporary directory containing the
package data.
}
\description{
This function generates a selection of test data for the "cran"
corpus, to allow functions to be run offline, without having to download the
large datasets otherwise required for the package to function.

Note that these data are randomly generated, and results will be generally
meaningless. They are generated solely to demonstrate how the package
functions, and are not intended to derive meaningful outputs.
}
\examples{
generate_pkgmatch_example_data ()
input <- "curl" # Name of a single installed package
# This generates a warning about "ollama" not being available:
suppressWarnings (
    pkgmatch_similar_pkgs (input, corpus = "cran")
)
}
\seealso{
Other utils: 
\code{\link{head.pkgmatch}()},
\code{\link{pkgmatch_browse}()},
\code{\link{pkgmatch_load_data}()},
\code{\link{pkgmatch_update_cache}()},
\code{\link{print.pkgmatch}()},
\code{\link{text_is_code}()}
}
\concept{utils}
