% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgmatch-methods.R
\name{head.pkgmatch}
\alias{head.pkgmatch}
\title{Head method for 'pkgmatch' objects}
\usage{
\method{head}{pkgmatch}(x, n = 5L, ...)
}
\arguments{
\item{x}{Object for which head is to be printed}

\item{n}{Number of rows of full \code{pkgmatch} object to be displayed}

\item{...}{Not used}
}
\value{
A (usually) smaller version of \code{x}, with all columns displayed.
}
\description{
Head method for 'pkgmatch' objects
}
\examples{
corpus <- "cran"
generate_pkgmatch_example_data (corpus = corpus)
input <- "Download open spatial data from NASA"
p <- pkgmatch_similar_pkgs (input, corpus = corpus)
head (p) # Shows first 5 rows of full `data.frame` object
}
\seealso{
Other utils:
\code{\link[=generate_pkgmatch_example_data]{generate_pkgmatch_example_data()}},
\code{\link[=pkgmatch_browse]{pkgmatch_browse()}},
\code{\link[=pkgmatch_load_data]{pkgmatch_load_data()}},
\code{\link[=pkgmatch_update_cache]{pkgmatch_update_cache()}},
\code{\link[=print.pkgmatch]{print.pkgmatch()}}
}
\concept{utils}
