% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{ollama_check}
\alias{ollama_check}
\title{Check ollama installation}
\usage{
ollama_check(sudo = is_docker_sudo())
}
\arguments{
\item{sudo}{Set to \code{TRUE} if ollama is running in docker with sudo
privileges.}
}
\value{
TRUE if everything works okay, otherwise the function will error
before returning, and issue an informative error message.
}
\description{
Performs the following checks:
\itemize{
\item Check that ollama is installed
\item Check that ollama is running
\item Check that ollama has the required models, and download if not
}

The required models are the Jina AI embeddings:
\url{https://ollama.com/jina/jina-embeddings-v2-base-en} for text
embeddings, and \url{https://ollama.com/unclemusclez/jina-embeddings-v2-base-code}
for code embeddings.

Note that the URL of a locally-running ollama instance is presumed by
default to be "127.0.0.1:11434". Other values can be set using the
\link{set_ollama_url} function.
}
\examples{
\dontrun{
chk <- ollama_check ()
}

}
\seealso{
Other ollama: 
\code{\link{get_ollama_url}()},
\code{\link{set_ollama_url}()}
}
\concept{ollama}
