% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{ollama_check}
\alias{ollama_check}
\title{Check that ollama is installed with required models, and download if not.}
\usage{
ollama_check(sudo = is_docker_sudo())
}
\arguments{
\item{sudo}{Set to \code{TRUE} if ollama is running in docker with sudo
privileges.}
}
\value{
TRUE if everything works okay, otherwise the function will error
before returning.
}
\description{
Note that the URL of a locally-running ollama instance is presumed by
default to be "127.0.0.1:11434". Other values can be set using the
\link{ollama_set_url} function.
}
\examples{
\dontrun{
chk <- ollama_check ()
}

}
\seealso{
Other ollama: 
\code{\link{get_ollama_url}()},
\code{\link{set_ollama_url}()}
}
\concept{ollama}
