% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgmatch-package.R
\docType{package}
\name{pkgmatch-package}
\alias{pkgmatch}
\alias{pkgmatch-package}
\title{pkgmatch: Find R Packages Matching Either Descriptions or Other R Packages}
\description{
Find R packages from CRAN, rOpenSci, or Bioconductor corpora. Packages can be matched to general text descriptions, to names of installed packages, or to local paths to entire source repositories. The package is used to list the most similar packages for each new submission to the rOpenSci software peer-review program (rOpenSci authors, 2026; \doi{10.5281/zenodo.18885936}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/pkgmatch/}
  \item \url{https://github.com/ropensci-review-tools/pkgmatch}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/pkgmatch/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

Authors:
\itemize{
  \item Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})
}

Other contributors:
\itemize{
  \item Davis Vaughan \email{davis@posit.co} [contributor]
}

}
\keyword{internal}
