% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm25.R
\name{pkgmatch_bm25_fn_calls}
\alias{pkgmatch_bm25_fn_calls}
\title{The "Best Matching 25" (BM25) ranking function for function calls}
\usage{
pkgmatch_bm25_fn_calls(path, corpus = NULL)
}
\arguments{
\item{path}{Local path to source code of an R package.}

\item{corpus}{One of "ropensci" or "cran"}
}
\value{
A \code{data.frame} of two columns:
\itemize{
\item "package" Naming the package from the specified corpus;
\item bm25 The "BM25" index value for the nominated packages, where high
values indicate greater overlap in term frequencies.
}
}
\description{
See \code{?pkgmatch_bm25} for details of BM25 ranks. This function
calculates "BM25" ranks from function-call frequencies between a local R
package and all packages in specified corpus. Values are thus higher for
packages with similar patterns of function calls, weighted by inverse
frequencies, so functions called infrequently across the entire corpus
contribute more than common functions.

Note that the results of this function are entirely different from
\link{pkgmatch_bm25} with \code{corpus = "ropensci-fns"} or \code{corpus = "bioc-fns"}. The latter return BM25 values from text descriptions of all
functions in all rOpenSci or BioConductor packages, whereas this function
returns BM25 values based on frequencies of function calls within packages.
}
\examples{
\dontrun{
u <- "https://cran.r-project.org/src/contrib/odbc_1.5.0.tar.gz"
path <- file.path (tempdir (), basename (u))
download.file (u, destfile = path)
bm25 <- pkgmatch_bm25_fn_calls (path)
}
}
\seealso{
Other bm25: 
\code{\link{pkgmatch_bm25}()}
}
\concept{bm25}
