% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{pkgmatch_browse}
\alias{pkgmatch_browse}
\title{Open web pages for \code{pkgmatch} results}
\usage{
pkgmatch_browse(p, n = NULL)
}
\arguments{
\item{p}{A \code{pkgmatch} object returned from either
\link{pkgmatch_similar_pkgs} or \link{pkgmatch_similar_fns}.}

\item{n}{Number of top-matching entries which should be opened. Defaults to
the value passed to the main functions.}
}
\value{
(Invisibly) A named vector of integers, with 0 for all pages able to
be successfully opened, and 1 otherwise.
}
\description{
Open web pages for \code{pkgmatch} results
}
\examples{
\dontrun{
input <- "genomics and transcriptomics sequence data"
p <- pkgmatch_similar_pkgs (input)
pkgmatch_browse (p) # Open main package pages on rOpenSci
p <- pkgmatch_similar_pkgs (input, corpus = "cran")
pkgmatch_browse (p) # Open main package pages on CRAN
p <- pkgmatch_similar_fns (input)
pkgmatch_browse (p) # Open pages for best-matching rOpenSci functions
}
}
\seealso{
Other utils: 
\code{\link{generate_pkgmatch_example_data}()},
\code{\link{head.pkgmatch}()},
\code{\link{pkgmatch_load_data}()},
\code{\link{pkgmatch_update_cache}()},
\code{\link{print.pkgmatch}()},
\code{\link{text_is_code}()}
}
\concept{utils}
