% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{pkgmatch_embeddings_from_pkgs}
\alias{pkgmatch_embeddings_from_pkgs}
\title{Return raw embeddings from package text and function definitions.}
\usage{
pkgmatch_embeddings_from_pkgs(packages = NULL, functions_only = FALSE)
}
\arguments{
\item{packages}{A vector of either names of installed packages, or local
paths to directories containing R packages.}

\item{functions_only}{If \code{TRUE}, calculate embeddings for function
descriptions only. This is intended to generate a separate set of embeddings
which can then be used to match plain-text queries of functions, rather than
entire packages.}
}
\value{
If \code{!functions_only}, a list of two matrices of embeddings: one for
the text descriptions of the specified packages, including individual
descriptions of all package functions, and one for the entire code base. For
\code{functions_only}, a single matrix of embeddings for all function
descriptions.
}
\description{
This function accepts a vector of either names of installed
packages, or paths to local source code directories, and calculates language
model (LM) embeddings for both text descriptions within the package
(documentation, including of functions), and for the entire code base.
Embeddings may also be calculating separately for all function descriptions.

The embeddings are currently retrieved from a local 'ollama' server
(\url{https://ollama.com}) running Jina AI embeddings
(\url{https://ollama.com/jina/jina-embeddings-v2-base-en} for text, and
\url{https://ollama.com/ordis/jina-embeddings-v2-base-code} for code).
}
\examples{
packages <- "curl"
emb_fns <- pkgmatch_embeddings_from_pkgs (packages, functions_only = TRUE)
colnames (emb_fns) # All functions the package
emb_pkg <- pkgmatch_embeddings_from_pkgs (packages, functions_only = FALSE)
names (emb_pkg)
colnames (emb_pkg$text_with_fns) # "curl"
}
\seealso{
Other embeddings: 
\code{\link{pkgmatch_embeddings_from_text}()}
}
\concept{embeddings}
