% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar-fns.R
\name{pkgmatch_similar_fns}
\alias{pkgmatch_similar_fns}
\title{Identify R functions best matching a given input string}
\usage{
pkgmatch_similar_fns(input, corpus = "ropensci", n = 5L, browse = FALSE)
}
\arguments{
\item{input}{A text string.}

\item{corpus}{One of "ropensci" or "bioc" (for BioConductor). It is not
possible to match functions again CRAN packages.}

\item{n}{When the result of this function is printed to screen, the top \code{n}
packages will be displayed.}

\item{browse}{If \code{TRUE}, automatically open webpages of the top \code{n} matches
in local browser.}
}
\value{
A modified \code{data.frame} object of class "pkgmatch". The \code{data.frame}
has 3 columns:
\enumerate{
\item "pkg_fn" with the name of the function in the form
"package::function";
\item "simil" with a similarity score between 0 and 1; and
\item "rank" as an integer index, with the highest rank of 1 as the first row.
}
The return object has a default \code{print} method which prints the names only
of the first 5 best matching functions; see \code{?print.pkgmatch} for details.
}
\description{
Function matching is only available for functions from the
corpora of rOpenSci or Bioconductor packages, and not for CRAN packages.
}
\examples{
corpus <- "ropensci"
generate_pkgmatch_example_data (corpus = corpus)
input <- "Package for pretty cli output"
p <- pkgmatch_similar_fns (input, corpus = corpus)
p # Default print method, lists 5 best matching functions
head (p) # Shows first 5 rows of full `data.frame` object
}
\seealso{
Other main:
\code{\link[=pkgmatch_similar_pkgs]{pkgmatch_similar_pkgs()}}
}
\concept{main}
