% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar-pkgs.R
\name{pkgmatch_similar_pkgs}
\alias{pkgmatch_similar_pkgs}
\title{Find R packages matching an input of either text or another package}
\usage{
pkgmatch_similar_pkgs(
  input,
  corpus = NULL,
  idfs = NULL,
  n = 5L,
  browse = FALSE
)
}
\arguments{
\item{input}{Either a text string, a path to local source code of an R
package, or the name of any installed R package.}

\item{corpus}{Must be specified as one of "ropensci", "cran", or "bioc" (for
BioConductor). If \code{idfs} parameter is not specified, data will be
automatically downloaded for the corpus specified by this parameter. The
function will then return the most similar package from the specified
corpus. Note that calculations will \code{corpus = "cran"} will generally take
longer, because the corpus is much larger.}

\item{idfs}{Inverse Document Frequency tables for a suite of packages,
generated from \link{pkgmatch_bm25}. If not provided, pre-generated IDF
tables will be downloaded and stored in a local cache directory.}

\item{n}{When the result of this function is printed to screen, the top \code{n}
packages will be displayed.}

\item{browse}{If \code{TRUE}, automatically open webpages of the top \code{n} matches
in local browser.}
}
\value{
A \code{data.frame} with a "package" column naming packages, and one or
more columns of package ranks in terms of text similarity and, if \code{input} is
an R package, of similarity in code structure.

The returned object has a default \code{print} method which prints the best 5
matches directly to the screen, yet returns information on all packages
within the specified corpus. This information is in the form of a
\code{data.frame}, with one column for the package name, and one or more
additional columns of integer ranks for each package. There is also a \code{head}
method to print the first few entries of these full data (default \code{n = 5}).
To see all data, use \code{as.data.frame()}. See the example below for how to
manipulate these objects.
}
\description{
This function accepts as \code{input} either a text description, the
name of a locally-installed package, or a path to a local directory containing an R package.
It ranks all R packages within the specified corpus in terms of how well they
match that input. The "corpus" argument can specify either \href{https://ropensci.org/packages}{rOpenSci's package suite},
\href{https://cran.r-project.org}{CRAN}, or
\href{https://bioconductor.org}{Bioconductor}.

Ranks are obtained from scores derived from \href{https://en.wikipedia.org/wiki/Okapi_BM25}{"Best Match 25" (BM25)} scores based on document
token frequencies.

Ranks are generally obtained by matching both for full package text from the
specified corpus, including all long-form documentation, and by matching
package descriptions only. The function returns a single rank derived by
combining individual ranks using the \href{https://plg.uwaterloo.ca/~gvcormac/cormacksigir09-rrf.pdf}{Reciprocal Rank Fusion (RRF) algorithm}.

Finally, all components of this function are locally cached for each call
(by the \pkg{memoise} package), so additional calls to this function with
the same \code{input} and \code{corpus} should be much faster than initial calls.
}
\note{
The first time this function is run without passing \code{idfs}, required
values will be automatically downloaded and stored in a locally persistent
cache directory. Especially for the "cran" corpus, this downloading may take
quite some time.
}
\examples{
# The following function simulates remote data in temporary directory, to
# enable package usage without downloading. Do not run for normal usage.
generate_pkgmatch_example_data ()

input <- "curl" # Name of a single installed package
p <- pkgmatch_similar_pkgs (input, corpus = "cran")
p # Default print method, lists 5 best matching packages
head (p) # Shows first 5 rows of full `data.frame` object
}
\seealso{
Other main: 
\code{\link{pkgmatch_similar_fns}()}
}
\concept{main}
