% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{pkgmatch_update_cache}
\alias{pkgmatch_update_cache}
\title{Update all locally-cached \code{pkgmatch} data to latest versions.}
\usage{
pkgmatch_update_cache()
}
\description{
This function forces all locally-cached data to be updated with
latest version of remote data provided on the latest release of GitHub
repository at
\url{https://github.com/ropensci-review-tools/pkgmatch/releases}.

Caching strategies are described in the "\emph{Data Caching and Updating}"
vignette, accessible either locally via
\code{vignette("data-caching-and-updating", package = "pkgmatch")}, or online at
\url{https://docs.ropensci.org/pkgmatch/articles/C_data-caching-and-updating.html}.
In short, locally-cached data used by this package are updated
by default every 30 days (with the vignette describing how to modify this
default behaviour). This function forces all locally-cached data to be
updated, regardless of update frequencies.
}
\examples{
\dontrun{
pkgmatch_update_cache ()
}
}
\seealso{
Other utils: 
\code{\link{head.pkgmatch}()},
\code{\link{pkgmatch_browse}()},
\code{\link{pkgmatch_load_data}()},
\code{\link{print.pkgmatch}()},
\code{\link{text_is_code}()}
}
\concept{utils}
