% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{pkgmatch_update_cache}
\alias{pkgmatch_update_cache}
\title{Update all locally-cached \code{pkgmatch} data to latest versions.}
\usage{
pkgmatch_update_cache()
}
\description{
Locally-cached data used by this package are generally updated every 30
days, or according to \code{options("pkgmatch.update_frequency")} if that is set.
This function forces all locally-cached data to be updated, regardless of
update frequencies. Remote data are provided on the latest release of GitHub
repository at
\url{https://github.com/ropensci-review-tools/pkgmatch/releases}.
}
\examples{
\dontrun{
pkgmatch_update_cache ()
}
}
\seealso{
Other utils: 
\code{\link{head.pkgmatch}()},
\code{\link{pkgmatch_browse}()},
\code{\link{pkgmatch_load_data}()},
\code{\link{print.pkgmatch}()},
\code{\link{text_is_code}()}
}
\concept{utils}
