% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-update.R
\name{pkgmatch_update_data}
\alias{pkgmatch_update_data}
\title{Update pkgmatch` data for both CRAN and rOpenSci packages on GitHub release}
\usage{
pkgmatch_update_data(upload = TRUE)
}
\arguments{
\item{upload}{If \code{TRUE}, upload updated results to GitHub release.}
}
\value{
Local path to directory containing updated results.
}
\description{
This function is intended for internal rOpenSci use only. Usage by any
unauthorized users will error and have no effect unless run with \code{upload = FALSE}, in which case updated data will be created in the sub-directory
"pkgmatch-results" of R's current temporary directory. This updating may
take a very long time!
}
\concept{data}
