% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgmatch-methods.R
\name{print.pkgmatch}
\alias{print.pkgmatch}
\title{Print method for 'pkgmatch' objects}
\usage{
\method{print}{pkgmatch}(x, ...)
}
\arguments{
\item{x}{Object to be printed}

\item{...}{Additional parameters passed to default 'print' method.}
}
\value{
The result of printing \code{x}, in form of either a single character
vector, or a named list of character vectors.
}
\description{
The main \code{pkgmatch} functions, \link{pkgmatch_similar_pkgs} and
\link{pkgmatch_similar_fns}, return \code{data.frame} objects of class
"pkgmatch". This class exists primarily to enable this print method, which
summarises by default the top 5 matching packages or functions. Objects can
be converted to standard \code{data.frame}s with \code{as.data.frame()}.
}
\examples{
\dontrun{
input <- "Download open spatial data from NASA"
p <- pkgmatch_similar_pkgs (input)
p # Default print method, lists 5 best matching packages
head (p) # Shows first 5 rows of full `data.frame` object
}
}
\seealso{
Other utils: 
\code{\link{generate_pkgmatch_example_data}()},
\code{\link{head.pkgmatch}()},
\code{\link{pkgmatch_browse}()},
\code{\link{pkgmatch_load_data}()},
\code{\link{pkgmatch_update_cache}()},
\code{\link{text_is_code}()}
}
\concept{utils}
